pyproject.toml
[tool.poetry]
name = "py-concurrent-consumer"
version = "0.1.0"
description = ""
authors = ["Usernamelesss <8vd2roma@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
kafka-python = "2.0.2"
retry = "0.9.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytype = "^2023.3.31"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
